# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/get_test": {
            "get": {
                "description": "test, return hello",
                "produces": [
                    "application/json"
                ],
                "summary": "test",
                "parameters": [
                    {
                        "type": "string",
                        "description": "apikey",
                        "name": "apikey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "description": "API接口健康检查，Get请求",
                "produces": [
                    "application/json"
                ],
                "summary": "健康检查",
                "responses": {
                    "10000": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/post_test": {
            "post": {
                "description": "test, return hello",
                "produces": [
                    "application/json"
                ],
                "summary": "test",
                "parameters": [
                    {
                        "description": "apikey",
                        "name": "apikey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}